Simplified Authentication Storage  - Renamed from `Db` to `Mapper`  - Added SM aliases for easier overriding (Mapper and Session)
