+++ config/module.config.php
@@ -40,14 +40,15 @@ return array(
     'service_manager' => array(
         'invokables' => array(
             'ZfcUser\Authentication\Adapter\Db' => 'ZfcUser\Authentication\Adapter\Db',
-            'ZfcUser\Authentication\Storage\Db' => 'ZfcUser\Authentication\Storage\Db',
             'ZfcUser\Form\Login'                => 'ZfcUser\Form\Login',
             'zfcuser_user_service'              => 'ZfcUser\Service\User',
+            'zfcuser_authentication_storage_backend_session' => 'Zend\Authentication\Storage\Session',
         ),
         'factories' => array(
             'zfcuser_module_options'                        => 'ZfcUser\Factory\ModuleOptionsFactory',
             'zfcuser_auth_service'                          => 'ZfcUser\Factory\AuthenticationServiceFactory',
             'ZfcUser\Authentication\Adapter\AdapterChain'   => 'ZfcUser\Authentication\Adapter\AdapterChainServiceFactory',
+            'ZfcUser\Authentication\Storage\Mapper'         => 'ZfcUser\Factory\Authentication\Storage\MapperFactory',
             'zfcuser_login_form'                            => 'ZfcUser\Factory\Form\LoginFormFactory',
             'zfcuser_register_form'                         => 'ZfcUser\Factory\Form\RegisterFormFactory',
             'zfcuser_change_password_form'                  => 'ZfcUser\Factory\Form\ChangePasswordFormFactory',
@@ -58,6 +59,8 @@ return array(
         'aliases' => array(
             'zfcuser_register_form_hydrator' => 'zfcuser_user_hydrator',
             'zfcuser_zend_db_adapter' => 'Zend\Db\Adapter\Adapter',
+            'zfcuser_authentication_storage' => 'ZfcUser\Authentication\Storage\Mapper',
+            'zfcuser_authentication_storage_backend' => 'zfcuser_authentication_storage_backend_session',
         ),
     ),
     'controller_plugins' => array(
